The ••• menu splits along the seams Kobalte drew
`NodeMenu.tsx` was 621 lines because a HAND-ROLLED panel genuinely is one
thing: being open, the list, the question, the dismissals and the message were
all the same forty lines of state. Adopting the primitive (PR #158) took the
behaviour out and left the seams showing, and this is the file catching up with
that.

  - `action.ts` — what an entry IS, to the panel that draws it. The seam, and
    the mistake it makes visible: `actions.ts` is a pure table with a unit test
    and it was importing a `.tsx` COMPONENT to learn what a row of itself looks
    like. Both depend on the description now; neither on the other.
  - `Dots.tsx` — the `•••` before anybody presses it. Not a menu concern at
    all: it is a COST decision that exists because a shut `DropdownMenu` is 140
    IntersectionObservers and 33 MB on this app's own roadmap.
  - `Panel.tsx` / `Confirm.tsx` — what is inside the open panel. After the
    primitive there is no state left in them but `asking`; they are Kobalte's
    Items wearing this app's classes, which is markup and nothing else.
  - `saying.ts` / `Said.tsx` — running a verb and what it had to say. Never the
    menu: the panel is GONE by the time most verbs answer, so this outlives the
    primitive and belongs to the row.

What is left in `NodeMenu.tsx` is the four things true of the menu as a menu:
what owns it, where it is drawn, what it hangs off on each of the two doors,
and where the caret goes. 621 lines to 300.

`saying.test.ts` is the split paying for itself: its four rules (nothing to
say, the ops layer's sentence verbatim, the throw worded from the label, the
line cleared before the next verb) were reachable only by driving a browser
while they lived inside a component.
